TOBAGO-1539: Summary: Using Servelt 3.0 for uploading files with <tc:file>

+++ tobago-example/tobago-example-demo/src/main/java/org/apache/myfaces/tobago/example/demo/UploadItem.java
@@ -29,10 +29,10 @@ public class UploadItem implements Serializable {
   private static final Logger LOG = LoggerFactory.getLogger(UploadItem.class);
 
   private String name;
-  private int size;
+  private long size;
   private String type;
 
-  public UploadItem(final String name, final int size, final String type) {
+  public UploadItem(final String name, final long size, final String type) {
     this.name = name;
     this.size = size;
     this.type = type;
@@ -46,11 +46,11 @@ public class UploadItem implements Serializable {
     this.name = name;
   }
 
-  public int getSize() {
+  public long getSize() {
     return size;
   }
 
-  public void setSize(final int size) {
+  public void setSize(final long size) {
     this.size = size;
   }
